+++ src/Minimizations/SequentialSubspaceMinimizerNoise.cpp
@@ -19,8 +19,8 @@
 #include "MinimizationExceptions.hpp"
 
 SequentialSubspaceMinimizerNoise::SequentialSubspaceMinimizerNoise(
-		const unsigned int _NormX,
-		const unsigned int _NormY,
+		const double _NormX,
+		const double _NormY,
 		const double _PowerX,
 		const double _PowerY,
 		const double _Delta,
